Add files via upload
Semester 2 Project - Via Canteen Booking System: This system was developed to support a multi-user environment. The front-end (GUI) was created using Scene Builder, while the client and server code were written in Java. To facilitate communication between the client and server, we utilized the RMI (Remote Method Invocation) protocol. The database was built using PostgreSQL and hosted on the ElephantSQL online database platform.